fix(clippy): remove unnecessary cast and add explicit lifetime in anomaly_detector
- lib.rs:1007: Remove redundant `as u32` cast since Soroban Vec::len()
  already returns u32
- test.rs:12: Add explicit `'_` lifetime to AnomalyDetectorContractClient
  in setup() return type to fix mismatched_lifetime_syntaxes lint

+++ contracts/anomaly_detector/src/lib.rs
@@ -1004,7 +1004,7 @@ impl AnomalyDetectorContract {
         weights: &Vec<u32>,
         feature_names: &Vec<String>,
     ) -> (u32, Vec<FeatureContribution>) {
-        let n = features.len().min(weights.len()) as u32;
+        let n = features.len().min(weights.len());
         let mut weighted_sum: u64 = 0;
         let mut total_weight: u64 = 0;
         let mut contributions = Vec::new(env);
